agent/scheduler: Use setsid in daemonize.
* This fixes a bug where CTRL-C'ing the _first_ command (aka the one
spinning up implictely the agent) would just terminate the agent. That's
because without setsid the new process (the agent) is part of the same
session and group and as such receives the CTRL-C signal. Call
ForkAndExec with Setsid set so that the new process creates a new
session and group.